Break out NoteDecoderTests
