+++ Tests/MusicXMLTests/DecoderTests.swift
@@ -11,30 +11,6 @@ import MusicXML
 
 class DecoderTests: XCTestCase {
 
-    func testPitch() {
-        let xml = """
-        <pitch>
-          <step>C</step>
-          <octave>4</octave>
-        </pitch>
-        """
-        let _ = try! XMLDecoder().decode(MusicXML.Pitch.self, from: xml.data(using: .utf8)!)
-    }
-
-    func testNote() {
-        let xml = """
-        <note>
-            <pitch>
-                <step>C</step>
-                <octave>4</octave>
-            </pitch>
-            <duration>4</duration>
-            <type>whole</type>
-        </note>
-        """
-        let _ = try! XMLDecoder().decode(MusicXML.Note.self, from: xml.data(using: .utf8)!)
-    }
-
     func testTime() {
         let xml = """
         <time>
